requirements.txt
# requirements.txt
google-genai
python-dotenv
duckduckgo_search
requests